Parser#initialize: read file when filename given
It seems natural to pass a filename to initialize, so if whatever we're
given is a file that exists, read its contents instead of using it as
string data directly.

(It occurs to me that we may want to duplicate the string otherwise, but
I'm unsure what other use cases might be present that I'm not thinking
of, so I'm leaving that out for now.)

Fixes #2

Note: Chosing to bump version to 0.2.2 because rubygems has a 0.2.1 from
somewhere that I couldn't find.

This change technically Copyright 2021 by David Lindes, and I agree to
license it under the same terms as the rest of this repository.