added new metadata api
This allows access to some additional information in the file, like the
date that it was reported to have been generated.

To accomodate this, QuickenParser::parser has been added, which is just
like the old QuickenParser::parse (which still exists, with unchanged
semantics), except that instead of returning the Parser#accounts array,
it returns the Parser object itself, which one can then call #accounts
and/or #metadata on as one pleases.